+++ app/src/features/admin/policies/detail/skeleton/PolicySkeleton.tsx
